pyproject.toml
[tool.poetry]
name = "re-bot"
version = "0.0.0"
description = ""
authors = ["Philipp Gayret <philipp@release-engineers.com>"]
readme = "README.md"
packages = [{ include = "rebot" }]
license = "SSPL-1.0"
classifiers = ["Private :: Do not Upload"]

[tool.poetry.scripts]
"rebot" = "rebot.main:main"

[tool.poetry.dependencies]
python = "^3.12"
slack-bolt = "^1.18.1"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
z3-solver = "^4.12.5.0"
sqlalchemy = "^2.0.25"
pyinvoker = { path = "../pyinvoker" }

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
